+++ protected/humhub/modules/notification/views/layouts/mail.php
@@ -19,9 +19,6 @@
                                 <!-- start content left -->
                                 <table width="100%" border="0" cellspacing="0" cellpadding="0" align="left">
 
-                              
-
-
                                     <!-- start content top-->
                                     <tr>
                                         <td valign="top" align="left">