+++ protected/humhub/modules/notification/views/mails/default.php
@@ -1,3 +1,12 @@
 <?php $this->beginContent('@notification/views/layouts/mail.php', $_params_); ?>
-    <?= $html; ?>
+<?= $html; ?>
+<br />
+<br />
+<?=
+\humhub\widgets\mails\MailButtonList::widget([
+    'buttons' => [
+        humhub\widgets\mails\MailButton::widget(['url' => $url, 'text' => Yii::t('ContentModule.notifications_mails', 'View Online')])
+    ]
+])
+?>
 <?php $this->endContent(); ?>